requirements.txt
RPi.Gpio
Flask